+++ tsend.c
@@ -129,7 +129,7 @@ again:
 		);
 	TSEND_ERR_CHECK("tsend_stream");
 	written+=n; 
-	if (n<len){ 
+	if (n<(int)len){ 
 		/* partial write */ 
 		buf+=n; 
 		len-=n;
